Add e2e tests for knocking into public knock rooms
