+++ cypress/e2e/spotlight/spotlight.spec.ts
@@ -23,35 +23,12 @@ import Loggable = Cypress.Loggable;
 import Timeoutable = Cypress.Timeoutable;
 import Withinable = Cypress.Withinable;
 import Shadow = Cypress.Shadow;
-
-enum Filter {
-    People = "people",
-    PublicRooms = "public_rooms",
-}
+import { Filter } from "../utils";
 
 declare global {
     // eslint-disable-next-line @typescript-eslint/no-namespace
     namespace Cypress {
         interface Chainable {
-            /**
-             * Opens the spotlight dialog
-             */
-            openSpotlightDialog(
-                options?: Partial<Loggable & Timeoutable & Withinable & Shadow>,
-            ): Chainable<JQuery<HTMLElement>>;
-            spotlightDialog(
-                options?: Partial<Loggable & Timeoutable & Withinable & Shadow>,
-            ): Chainable<JQuery<HTMLElement>>;
-            spotlightFilter(
-                filter: Filter | null,
-                options?: Partial<Loggable & Timeoutable & Withinable & Shadow>,
-            ): Chainable<JQuery<HTMLElement>>;
-            spotlightSearch(
-                options?: Partial<Loggable & Timeoutable & Withinable & Shadow>,
-            ): Chainable<JQuery<HTMLElement>>;
-            spotlightResults(
-                options?: Partial<Loggable & Timeoutable & Withinable & Shadow>,
-            ): Chainable<JQuery<HTMLElement>>;
             roomHeaderName(
                 options?: Partial<Loggable & Timeoutable & Withinable & Shadow>,
             ): Chainable<JQuery<HTMLElement>>;
@@ -59,57 +36,6 @@ declare global {
         }
     }
 }
-
-Cypress.Commands.add(
-    "openSpotlightDialog",
-    (options?: Partial<Loggable & Timeoutable & Withinable & Shadow>): Chainable<JQuery<HTMLElement>> => {
-        cy.get(".mx_RoomSearch_spotlightTrigger", options).click({ force: true });
-        return cy.spotlightDialog(options);
-    },
-);
-
-Cypress.Commands.add(
-    "spotlightDialog",
-    (options?: Partial<Loggable & Timeoutable & Withinable & Shadow>): Chainable<JQuery<HTMLElement>> => {
-        return cy.get('[role=dialog][aria-label="Search Dialog"]', options);
-    },
-);
-
-Cypress.Commands.add(
-    "spotlightFilter",
-    (
-        filter: Filter | null,
-        options?: Partial<Loggable & Timeoutable & Withinable & Shadow>,
-    ): Chainable<JQuery<HTMLElement>> => {
-        let selector: string;
-        switch (filter) {
-            case Filter.People:
-                selector = "#mx_SpotlightDialog_button_startChat";
-                break;
-            case Filter.PublicRooms:
-                selector = "#mx_SpotlightDialog_button_explorePublicRooms";
-                break;
-            default:
-                selector = ".mx_SpotlightDialog_filter";
-                break;
-        }
-        return cy.get(selector, options).click();
-    },
-);
-
-Cypress.Commands.add(
-    "spotlightSearch",
-    (options?: Partial<Loggable & Timeoutable & Withinable & Shadow>): Chainable<JQuery<HTMLElement>> => {
-        return cy.get(".mx_SpotlightDialog_searchBox", options).findByRole("textbox", { name: "Search" });
-    },
-);
-
-Cypress.Commands.add(
-    "spotlightResults",
-    (options?: Partial<Loggable & Timeoutable & Withinable & Shadow>): Chainable<JQuery<HTMLElement>> => {
-        return cy.get(".mx_SpotlightDialog_section.mx_SpotlightDialog_results .mx_SpotlightDialog_option", options);
-    },
-);
 
 Cypress.Commands.add(
     "roomHeaderName",
