+++ cypress/e2e/utils.ts
@@ -15,7 +15,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { MatrixClient, MatrixEvent, Room } from "matrix-js-sdk/src/matrix";
+import type { MatrixClient, MatrixEvent, Room } from "matrix-js-sdk/src/matrix";
+
+export enum Filter {
+    People = "people",
+    PublicRooms = "public_rooms",
+}
 
 export function waitForRoom(
     win: Cypress.AUTWindow,
